+++ NEWS.md
@@ -25,6 +25,10 @@ for the time being.**
 *   Extended the `CLEAR` and `NEW` commands so that they reset various
     properties of the machine, not just variables.
 
+*   Modified the `RUN` command to issue a `CLEAR` upfront.  Maintaining the
+    previous state of the machine was just too confusing and error-prone due
+    to side-effects causing execution failures.
+
 ## Changes in version 0.7.0
 
 **Released on 2021-07-03.**